+++ package.json
@@ -7,7 +7,6 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "bootstrap": "^5.2.0-beta1",
     "date-fns": "^2.28.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
