--- src/Components/Header/Header.css
@@ -1,22 +1,0 @@
-header{
-	height:280px;
-	color: #fff;
-	text-align: center;
-	/*border: 1px solid #fff;*/
-}
-
-.contaner-banner{
-	border: 1px solid #fff;
-}
-
-header h1{
-	font-size: 100px;
-	color: #039bd9;
-	font-family: 'Poppins', sans-serif;
-}
-
-@media(max-width: 400px){
-	.container-banner h1{
-		font-size:90px;
-	}
-}
