--- src/Components/Navbar/Navbar.css
@@ -1,48 +1,0 @@
-.container-nav{
-	padding: 1em;
-	margin: 2em auto 5em;
-	max-width: 1000px;
-	display: flex;
-    justify-content:space-between;
-}
-
-.container-nav img{
-	width: 130px;
-	height: 100%;
-}
-
-.container-nav a{
-	text-decoration: none;
-	color: #039bd9;
-	font-family: 'Poppins', sans-serif;
-}
-
-#nav-hr{
-	border: none;
-    height: 1px;
-	color: #413F42;
-	background-color: #413F42;
-}
-
-.nav-logo{
-	display: flex;
-    justify-content:flex-start;
-    align-items: center;
-}
-
-.nav-links{
-    display: flex;
-    justify-content: right;
-    align-items: center;
-} 
-
-.nav-links{
-    display: flex;
-    justify-content: right;
-    align-items: center;
-} 
-
-.nav-links li{
-    display: inline;
-    padding: 1em;  
-}
