--- src/Components/Post/Post.css
@@ -1,22 +1,0 @@
-.post-container-c{
-	margin: 2em auto;
-	max-width: 1000px;
-	text-align: justify;
-	padding: 2em;
-}
-
-.post-container-c img{
-	width: 100%;
-	height: 350px;
-	object-fit: cover;
-}
-.post-container-c a{
-/*	color: #000;*/
-}
-.post-container-c a:hover{
-	color: orange;
-	text-decoration: underline;
-}
-.post-container-c h2{
-	margin: 1em 0;
-}
