--- src/Services.jsx
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-function Services(){
-
-
-}
-export default Services;
