+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,49 @@
+.container-nav {
+	padding: 0.4em;
+	margin: 1em auto 0;
+	max-width: 1000px;
+	display: flex;
+	justify-content: space-between;
+}
+
+.container-nav img {
+	width: 130px;
+	height: 100%;
+	margin-left: 0.4em;
+}
+
+.container-nav a {
+	text-decoration: none;
+	color: #039bd9;
+	font-family: 'Poppins', sans-serif;
+	font-weight: 500;
+}
+
+.nav-logo {
+	display: flex;
+	justify-content: flex-start;
+	align-items: center;
+}
+
+.nav-links {
+	display: flex;
+	justify-content: right;
+	align-items: center;
+}
+
+.nav-links {
+	display: flex;
+	justify-content: right;
+	align-items: center;
+}
+
+.nav-links li {
+	display: inline;
+	padding: 1em;
+}
+
+@media(max-width: 400px) {
+	.container-nav img {
+		width: 110px;
+	}
+}
